Refactor CSS
