+++ core/services/__init__.py
@@ -1,9 +1,9 @@
 """Core services package."""
 
 from .paper_service import PaperService
-from .paper_summarization_service import PaperSummarizationService
 from .star_service import StarService
 from .stream_service import StreamService
+from .summarization_service import PaperSummarizationService
 
 __all__ = [
     "PaperService",
