--- static/css/base.css
@@ -1,90 +1,0 @@
-/* Reset and base styles */
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-body {
-    font-family: 'Consolas', 'Monaco', 'Courier New', monospace;
-    line-height: 1.6;
-    color: #333;
-    background-color: #f8f9fa;
-    height: 100vh;
-    display: flex;
-    flex-direction: column;
-    overflow: hidden;
-}
-
-/* Korean text font */
-.korean-text {
-    font-family: 'Consolas', 'Malgun Gothic', '맑은 고딕', 'Monaco', 'Courier New', monospace;
-}
-
-/* Container */
-.container {
-    width: 100%;
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    overflow: hidden;
-}
-
-/* Fixed header */
-.header {
-    flex-shrink: 0;
-    padding: 0.5rem 0.5rem 0;
-    background: #f8f9fa;
-    z-index: 100;
-    font-family: 'Consolas', 'Malgun Gothic', '맑은 고딕', 'Monaco', 'Courier New', monospace;
-}
-
-/* Scrollable content area */
-.content {
-    flex: 1;
-    overflow-y: auto;
-    min-height: 0; /* Allow content to shrink properly */
-}
-
-/* Custom scrollbar styling */
-.content::-webkit-scrollbar {
-    width: 8px;
-}
-
-.content::-webkit-scrollbar-track {
-    background: #f1f1f1;
-    border-radius: 4px;
-}
-
-.content::-webkit-scrollbar-thumb {
-    background: #c1c1c1;
-    border-radius: 4px;
-}
-
-.content::-webkit-scrollbar-thumb:hover {
-    background: #a8a8a8;
-}
-
-/* Footer - Fixed at bottom */
-.footer {
-    flex-shrink: 0;
-    background: white;
-    border-top: 1px solid #e1e5e9;
-    padding: 0.5rem;
-    text-align: right;
-    color: #6c757d;
-    font-size: 0.9rem;
-    font-family: 'Consolas', 'Malgun Gothic', '맑은 고딕', 'Monaco', 'Courier New', monospace;
-    width: 100%;
-}
-
-/* Main content */
-.main {
-    padding: 1rem 0;
-    min-height: 0; /* Allow main to shrink properly */
-}
-
-/* Hidden class */
-.hidden {
-    display: none !important;
-}
