--- static/css/modal.css
@@ -1,209 +1,0 @@
-/* Modal styles for TheArk */
-
-.modal-overlay {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background: rgba(0, 0, 0, 0.5);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    z-index: 1000;
-}
-
-.modal-content {
-    background: white;
-    border-radius: 8px;
-    box-shadow: 0 4px 20px rgba(0, 0, 0, 0.3);
-    padding: 8px;
-    max-width: 90vw;
-    width: 800px;
-    max-height: 90vh;
-    display: flex;
-    flex-direction: column;
-}
-
-.modal-header {
-    flex-shrink: 0;
-    padding: 0.5rem;
-    border-bottom: 1px solid #eee;
-    background: #fff;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.modal-title {
-    font-size: 1.2rem;
-    font-weight: bold;
-    margin: 0;
-    flex: 1;
-}
-
-.modal-close {
-    background: none;
-    border: none;
-    font-size: 1.5rem;
-    cursor: pointer;
-    padding: 0.25rem;
-    color: #666;
-    transition: color 0.2s ease;
-}
-
-.modal-close:hover {
-    color: #333;
-}
-
-.modal-body {
-    flex: 1;
-    max-height: 60vh;
-    overflow-y: auto;
-    padding: 1rem;
-    background: #fff;
-}
-
-.modal-footer {
-    flex-shrink: 0;
-    padding: 0.5rem 0.25rem 0.25rem 0.25rem;
-    border-top: 1px solid #eee;
-    text-align: right;
-}
-
-.modal-tag-container {
-    position: relative;
-    width: 0;
-    height: 0;
-}
-
-.modal-star-button {
-    position: absolute;
-    top: -40px;
-    right: 10px;
-    background: none;
-    border: none;
-    cursor: pointer;
-    font-size: 1.2em;
-    padding: 0.2em;
-    min-width: 1.5em;
-    text-align: center;
-    z-index: 10;
-}
-
-.modal-star-button:hover {
-    transform: scale(1.1);
-}
-
-/* Summary content styles */
-.summary-content {
-    line-height: 1.6;
-}
-
-.summary-section {
-    margin-bottom: 1.5rem;
-}
-
-.summary-section h3 {
-    margin: 0 0 0.5rem 0;
-    color: #333;
-    font-size: 1.1rem;
-}
-
-.summary-section p {
-    margin: 0;
-    color: #555;
-}
-
-/* Model info styles */
-.model-info {
-    font-size: 0.9em;
-    color: #666;
-    margin: 0;
-}
-
-.model-info p {
-    margin: 0;
-}
-
-/* Relevance tag styles */
-.relevance-tag-container {
-    position: absolute;
-    top: -30px;
-    left: 10px;
-    z-index: 10;
-}
-
-.relevance-tag {
-    background: #007bff;
-    color: white;
-    padding: 0.2rem 0.5rem;
-    border-radius: 4px;
-    font-size: 0.8rem;
-    font-weight: bold;
-}
-
-.relevance-tag.score-0 {
-    background: #6c757d;
-}
-
-.relevance-tag.score-1 { background: #dc3545; }
-.relevance-tag.score-2 { background: #fd7e14; }
-.relevance-tag.score-3 { background: #ffc107; }
-.relevance-tag.score-4 { background: #28a745; }
-.relevance-tag.score-5 { background: #20c997; }
-.relevance-tag.score-6 { background: #17a2b8; }
-.relevance-tag.score-7 { background: #6f42c1; }
-.relevance-tag.score-8 { background: #e83e8c; }
-.relevance-tag.score-9 { background: #fd7e14; }
-.relevance-tag.score-10 { background: #28a745; }
-
-/* Loading modal styles */
-.loading-modal {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(0, 0, 0, 0.5);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    z-index: 10000;
-}
-
-.loading-modal-content {
-    background-color: white;
-    padding: 2rem;
-    border-radius: 8px;
-    box-shadow: 0 4px 20px rgba(0, 0, 0, 0.3);
-    text-align: center;
-    max-width: 300px;
-}
-
-.loading-spinner {
-    font-size: 2rem;
-    margin-bottom: 1rem;
-    animation: spin 1s linear infinite;
-}
-
-.loading-text {
-    font-size: 1rem;
-    color: #666;
-}
-
-@keyframes spin {
-    from { transform: rotate(0deg); }
-    to { transform: rotate(360deg); }
-}
-
-/* Loading placeholder for overview */
-.loading-placeholder {
-    color: #999;
-    font-style: italic;
-    cursor: pointer;
-}
-
-.loading-placeholder:hover {
-    color: #666;
-}
