+++ templates/index.html
@@ -5,13 +5,9 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>TheArk - Paper Management</title>
-    <!-- Modular CSS files -->
-    <link rel="stylesheet" href="/static/css/base.css">
-    <link rel="stylesheet" href="/static/css/components.css">
-    <link rel="stylesheet" href="/static/css/papers.css">
-    <link rel="stylesheet" href="/static/css/star-button.css">
-    <link rel="stylesheet" href="/static/css/modal.css">
-    <link rel="stylesheet" href="/static/css/responsive.css">
+    <!-- Refactored CSS -->
+    <link rel="stylesheet" href="/static/css/variables.css">
+    <link rel="stylesheet" href="/static/css/main.css">
     <link rel="icon" href="/static/favicon.ico" type="image/x-icon">
 </head>
